refactor: split evaluator.c into modular detector files
Split the monolithic evaluator.c (764 lines) into a modular structure for improved
maintainability and navigation:

- Created src/detectors/ directory with 10 individual detector files:
  - royal_flush.c, straight_flush.c, four_of_a_kind.c, full_house.c
  - flush.c, straight.c, three_of_a_kind.c, two_pair.c
  - one_pair.c, high_card.c
- Created src/helpers.c with shared helper functions:
  - is_flush(), is_straight(), rank_counts(), rank_compare_desc()
- Updated src/evaluator.c to contain only poker_errno and evaluate_hand()
- Updated Makefile to compile new structure:
  - Added DETECTOR_SRC and DETECTOR_OBJ variables
  - Updated clean and coverage targets for detector subdirectory
  - Maintained backward compatibility with OBJ pattern for test suite
- Updated README with new project structure documentation
- All tests pass (19/19 test executables verified)

Benefits:
- Modularity: Each detector in its own file (50-100 lines)
- Separation of Concerns: Helpers separate from detectors
- Maintainability: Smaller files easier to understand and modify
- Navigation: Easy to locate specific hand detection logic

Closes #97
