+++ src/evaluator.c
@@ -1,780 +1,11 @@
-/* evaluator.c - Hand evaluation logic */
+/* evaluator.c - Main hand evaluation orchestration */
 
 #include "../include/poker.h"
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
 
 /* Global error indicator - initialized to POKER_EOK (0) */
 int poker_errno = 0;
 
-/* Static helper: Compare ranks in descending order for qsort
- * Uses conditional logic instead of subtraction to avoid integer overflow
- * Safety: Subtraction-based comparators can overflow when (b - a) > INT_MAX
- */
-static int rank_compare_desc(const void* a, const void* b) {
-    Rank rank_a = *(Rank*)a;
-    Rank rank_b = *(Rank*)b;
-
-    /* Descending order: larger ranks come first */
-    if (rank_b > rank_a) {
-        return 1;   /* b is greater, so b comes first */
-    } else if (rank_b < rank_a) {
-        return -1;  /* a is greater, so a comes first */
-    } else {
-        return 0;   /* Equal ranks */
-    }
-}
-
-/**
- * @brief Check if cards form a straight
- *
- * Detects straights in a HAND_SIZE-card hand, including special handling for
- * wheel straight (A-2-3-4-5). Sorts ranks internally before checking.
- *
- * @param cards Array of cards to check
- * @param len Number of cards (must be HAND_SIZE)
- * @param out_high_card Optional pointer to receive high card rank
- * @return 1 if straight detected, 0 otherwise
- */
-int is_straight(const Card* const cards, const size_t len, Rank* const out_high_card) {
-    if (len != HAND_SIZE) {
-        return 0;
-    }
-
-    /* Extract ranks into array */
-    Rank ranks[HAND_SIZE];
-    for (size_t i = 0; i < HAND_SIZE; i++) {
-        ranks[i] = (Rank)cards[i].rank;
-    }
-
-    /* Sort ranks in descending order */
-    qsort(ranks, HAND_SIZE, sizeof(Rank), rank_compare_desc);
-
-    /* Check for wheel straight: A-5-4-3-2 (sorted descending) */
-    if (ranks[0] == RANK_ACE && ranks[1] == RANK_FIVE &&
-        ranks[2] == RANK_FOUR && ranks[3] == RANK_THREE &&
-        ranks[4] == RANK_TWO) {
-        if (out_high_card != NULL) {
-            *out_high_card = RANK_FIVE;  /* Wheel: high card is 5 */
-        }
-        return 1;
-    }
-
-    /* Check for regular straight: each rank = previous - 1 */
-    for (size_t i = 1; i < HAND_SIZE; i++) {
-        if (ranks[i] != ranks[i-1] - 1) {
-            return 0;
-        }
-    }
-
-    /* Regular straight found */
-    if (out_high_card != NULL) {
-        *out_high_card = ranks[0];  /* Highest rank is first after sort */
-    }
-    return 1;
-}
-
-/**
- * @brief Check if all cards are the same suit
- *
- * Validates that all cards in the array have the same suit, which is required
- * for flush detection in poker hand evaluation.
- *
- * @param cards Array of cards to check
- * @param len Number of cards (must be HAND_SIZE for valid poker hand)
- * @return 1 if all cards have same suit, 0 otherwise
- */
-int is_flush(const Card* const cards, const size_t len) {
-    /* Validate input parameters */
-    if (cards == NULL || len != HAND_SIZE) {
-        return 0;
-    }
-
-    /* Get suit of first card */
-    uint8_t first_suit = cards[0].suit;
-
-    /* Check if all other cards have same suit */
-    for (size_t i = 1; i < len; i++) {
-        if (cards[i].suit != first_suit) {
-            return 0;
-        }
-    }
-
-    return 1;
-}
-
-/**
- * @brief Count frequency of each rank
- *
- * Initializes counts array to zero, then iterates through cards and
- * increments the count for each card's rank. The counts array is indexed
- * by Rank values (2-14), so it must be at least RANK_ARRAY_SIZE elements.
- *
- * Validates all input parameters and performs bounds checking on rank values
- * to prevent buffer overruns and undefined behavior.
- *
- * @param cards Array of cards to count (can be NULL if len is 0)
- * @param len Number of cards in array
- * @param counts Output array[RANK_ARRAY_SIZE] to receive counts (indexed by Rank, cannot be NULL)
- */
-void rank_counts(const Card* const cards, const size_t len, int* const counts) {
-    /* Validate counts pointer */
-    if (counts == NULL) {
-        return;
-    }
-
-    /* Initialize all counts to zero */
-    memset(counts, 0, RANK_ARRAY_SIZE * sizeof(int));
-
-    /* Handle empty or NULL cards array */
-    if (cards == NULL || len == 0) {
-        return;
-    }
-
-    /* Count occurrences of each rank with bounds checking */
-    for (size_t i = 0; i < len; i++) {
-        uint8_t rank = cards[i].rank;
-        /* Validate rank is within bounds (2-14) */
-        if (rank >= RANK_TWO && rank <= RANK_ACE) {
-            counts[rank]++;
-        }
-    }
-}
-
-/**
- * @brief Detect four of a kind
- *
- * Detects if the hand contains four cards of the same rank (four of a kind).
- * Returns tiebreakers in the format: [quad_rank, kicker].
- *
- * The function validates all input parameters and handles the optional counts
- * parameter by computing counts internally if NULL is provided.
- *
- * @param cards Array of exactly HAND_SIZE cards
- * @param len Must be HAND_SIZE
- * @param counts Optional pre-computed rank counts (can be NULL)
- * @param out_tiebreakers Output array for tiebreaker ranks
- * @param out_num_tiebreakers Pointer to receive count of tiebreakers
- * @return 1 if four of a kind, 0 otherwise
- */
-int detect_four_of_a_kind(const Card* const cards, const size_t len,
-                           const int* const counts,
-                           Rank* const out_tiebreakers,
-                           size_t* const out_num_tiebreakers) {
-    /* Validate input parameters */
-    if (cards == NULL || len != HAND_SIZE || out_tiebreakers == NULL || out_num_tiebreakers == NULL) {
-        return 0;
-    }
-
-    /* Use provided counts or compute locally */
-    int local_counts[RANK_ARRAY_SIZE];
-    const int* rank_count_array;
-
-    if (counts == NULL) {
-        rank_counts(cards, len, local_counts);
-        rank_count_array = local_counts;
-    } else {
-        rank_count_array = counts;
-    }
-
-    /* Find the quad rank (count == 4) */
-    Rank quad_rank = 0;
-    for (int rank = RANK_TWO; rank <= RANK_ACE; rank++) {
-        if (rank_count_array[rank] == 4) {
-            quad_rank = (Rank)rank;
-            break;
-        }
-    }
-
-    /* No quad found */
-    if (quad_rank == 0) {
-        return 0;
-    }
-
-    /* Find the kicker (count == 1) */
-    Rank kicker = 0;
-    for (int rank = RANK_TWO; rank <= RANK_ACE; rank++) {
-        if (rank_count_array[rank] == 1) {
-            kicker = (Rank)rank;
-            break;
-        }
-    }
-
-    /* Write tiebreakers: [quad_rank, kicker] */
-    out_tiebreakers[0] = quad_rank;
-    out_tiebreakers[1] = kicker;
-    *out_num_tiebreakers = 2;
-
-    return 1;
-}
-
-/**
- * @brief Detect straight flush
- *
- * Detects if the hand is a straight flush (HAND_SIZE sequential suited cards).
- * Combines is_flush() and is_straight() checks. Returns the high card
- * for tiebreakers (RANK_FIVE for wheel straight flush).
- *
- * @param cards Array of exactly HAND_SIZE cards
- * @param len Must be HAND_SIZE
- * @param out_high_card Pointer to receive high card rank (can be NULL)
- * @return 1 if straight flush, 0 otherwise
- */
-int detect_straight_flush(const Card* const cards, const size_t len, Rank* const out_high_card) {
-    /* Validate input length */
-    if (len != HAND_SIZE) {
-        return 0;
-    }
-
-    /* Check if all cards are the same suit (flush) */
-    if (!is_flush(cards, len)) {
-        return 0;
-    }
-
-    /* Check if cards form a straight */
-    if (!is_straight(cards, len, out_high_card)) {
-        return 0;
-    }
-
-    /* Both flush and straight confirmed - it's a straight flush */
-    return 1;
-}
-
-/**
- * @brief Detect royal flush
- *
- * Detects if the hand is a royal flush (10-J-Q-K-A all same suit).
- * Uses is_flush() to verify all cards have the same suit, then checks
- * for the exact ranks TEN, JACK, QUEEN, KING, ACE.
- *
- * Royal flush is the strongest poker hand and requires no tiebreakers.
- *
- * @param cards Array of exactly HAND_SIZE cards
- * @param len Must be HAND_SIZE
- * @return 1 if royal flush, 0 otherwise
- */
-int detect_royal_flush(const Card* const cards, const size_t len) {
-    /* Validate input length */
-    if (len != HAND_SIZE) {
-        return 0;
-    }
-
-    /* Check if all cards are the same suit (flush) */
-    if (!is_flush(cards, len)) {
-        return 0;
-    }
-
-    /* Track which royal ranks we've found */
-    int has_ten = 0;
-    int has_jack = 0;
-    int has_queen = 0;
-    int has_king = 0;
-    int has_ace = 0;
-
-    /* Check each card for royal ranks */
-    for (size_t i = 0; i < len; i++) {
-        Rank rank = (Rank)cards[i].rank;
-        if (rank == RANK_TEN) {
-            has_ten = 1;
-        } else if (rank == RANK_JACK) {
-            has_jack = 1;
-        } else if (rank == RANK_QUEEN) {
-            has_queen = 1;
-        } else if (rank == RANK_KING) {
-            has_king = 1;
-        } else if (rank == RANK_ACE) {
-            has_ace = 1;
-        } else {
-            /* Non-royal rank found - cannot be royal flush */
-            return 0;
-        }
-    }
-
-    /* Verify all royal ranks are present */
-    if (has_ten && has_jack && has_queen && has_king && has_ace) {
-        return 1;
-    }
-
-    return 0;
-}
-
-/**
- * @brief Detect full house
- *
- * Detects if the hand contains a full house (three of a kind + pair).
- * Returns tiebreakers in the format: [trip_rank, pair_rank].
- *
- * The function validates all input parameters and handles the optional counts
- * parameter by computing counts internally if NULL is provided.
- *
- * @param cards Array of exactly HAND_SIZE cards
- * @param len Must be HAND_SIZE
- * @param counts Optional pre-computed rank counts (can be NULL)
- * @param out_tiebreakers Output array for tiebreaker ranks
- * @param out_num_tiebreakers Pointer to receive count of tiebreakers
- * @return 1 if full house, 0 otherwise
- */
-int detect_full_house(const Card* const cards, const size_t len,
-                      const int* const counts,
-                      Rank* const out_tiebreakers,
-                      size_t* const out_num_tiebreakers) {
-    /* Validate input parameters */
-    if (cards == NULL || len != HAND_SIZE || out_tiebreakers == NULL || out_num_tiebreakers == NULL) {
-        return 0;
-    }
-
-    /* Use provided counts or compute locally */
-    int local_counts[RANK_ARRAY_SIZE];
-    const int* rank_count_array;
-
-    if (counts == NULL) {
-        rank_counts(cards, len, local_counts);
-        rank_count_array = local_counts;
-    } else {
-        rank_count_array = counts;
-    }
-
-    /* Find the trip rank (count == 3) */
-    Rank trip_rank = 0;
-    for (int rank = RANK_TWO; rank <= RANK_ACE; rank++) {
-        if (rank_count_array[rank] == 3) {
-            trip_rank = (Rank)rank;
-            break;
-        }
-    }
-
-    /* No trip found */
-    if (trip_rank == 0) {
-        return 0;
-    }
-
-    /* Find the pair rank (count == 2) */
-    Rank pair_rank = 0;
-    for (int rank = RANK_TWO; rank <= RANK_ACE; rank++) {
-        if (rank_count_array[rank] == 2) {
-            pair_rank = (Rank)rank;
-            break;
-        }
-    }
-
-    /* No pair found */
-    if (pair_rank == 0) {
-        return 0;
-    }
-
-    /* Write tiebreakers: [trip_rank, pair_rank] */
-    out_tiebreakers[0] = trip_rank;
-    out_tiebreakers[1] = pair_rank;
-    *out_num_tiebreakers = 2;
-
-    return 1;
-}
-
-/**
- * @brief Detect three of a kind (no pair)
- *
- * Detects if the hand contains three cards of the same rank without a pair
- * (excludes full houses). Returns tiebreakers in the format: [trip_rank, kicker1, kicker2]
- * where kickers are sorted in descending order.
- *
- * The function validates all input parameters and handles the optional counts
- * parameter by computing counts internally if NULL is provided.
- *
- * @param cards Array of exactly HAND_SIZE cards
- * @param len Must be HAND_SIZE
- * @param counts Optional pre-computed rank counts (can be NULL)
- * @param out_tiebreakers Output array for tiebreaker ranks
- * @param out_num_tiebreakers Pointer to receive count of tiebreakers
- * @return 1 if three of a kind, 0 otherwise
- */
-int detect_three_of_a_kind(const Card* const cards, const size_t len,
-                            const int* const counts,
-                            Rank* const out_tiebreakers,
-                            size_t* const out_num_tiebreakers) {
-    /* Validate input parameters */
-    if (cards == NULL || len != HAND_SIZE || out_tiebreakers == NULL || out_num_tiebreakers == NULL) {
-        return 0;
-    }
-
-    /* Use provided counts or compute locally */
-    int local_counts[RANK_ARRAY_SIZE];
-    const int* rank_count_array;
-
-    if (counts == NULL) {
-        rank_counts(cards, len, local_counts);
-        rank_count_array = local_counts;
-    } else {
-        rank_count_array = counts;
-    }
-
-    /* Find the trip rank (count == 3) */
-    Rank trip_rank = 0;
-    for (int rank = RANK_TWO; rank <= RANK_ACE; rank++) {
-        if (rank_count_array[rank] == 3) {
-            trip_rank = (Rank)rank;
-            break;
-        }
-    }
-
-    /* No trip found */
-    if (trip_rank == 0) {
-        return 0;
-    }
-
-    /* Check if there's a pair (which would make it a full house) */
-    for (int rank = RANK_TWO; rank <= RANK_ACE; rank++) {
-        if (rank_count_array[rank] == 2) {
-            /* Full house detected - return 0 */
-            return 0;
-        }
-    }
-
-    /* Find exactly 2 kickers (count == 1) */
-    Rank kickers[2];
-    int kicker_count = 0;
-
-    for (int rank = RANK_TWO; rank <= RANK_ACE; rank++) {
-        if (rank_count_array[rank] == 1) {
-            kickers[kicker_count++] = (Rank)rank;
-            if (kicker_count == 2) {
-                break;
-            }
-        }
-    }
-
-    /* Should have exactly 2 kickers */
-    if (kicker_count != 2) {
-        return 0;
-    }
-
-    /* Sort kickers in descending order */
-    if (kickers[0] < kickers[1]) {
-        Rank temp = kickers[0];
-        kickers[0] = kickers[1];
-        kickers[1] = temp;
-    }
-
-    /* Write tiebreakers: [trip_rank, kicker1, kicker2] */
-    out_tiebreakers[0] = trip_rank;
-    out_tiebreakers[1] = kickers[0];
-    out_tiebreakers[2] = kickers[1];
-    *out_num_tiebreakers = 3;
-
-    return 1;
-}
-
-/**
- * @brief Detect two pair
- *
- * Detects if the hand contains exactly two pairs (2 ranks with 2 cards each).
- * Returns tiebreakers in the format: [high_pair, low_pair, kicker]
- * where pairs are sorted in descending order.
- *
- * The function validates all input parameters and handles the optional counts
- * parameter by computing counts internally if NULL is provided. It excludes
- * full houses (trips with a pair) and quads.
- *
- * @param cards Array of exactly HAND_SIZE cards
- * @param len Must be HAND_SIZE
- * @param counts Optional pre-computed rank counts (can be NULL)
- * @param out_tiebreakers Output array for tiebreaker ranks
- * @param out_num_tiebreakers Pointer to receive count of tiebreakers
- * @return 1 if two pair, 0 otherwise
- */
-int detect_two_pair(const Card* const cards, const size_t len,
-                    const int* const counts,
-                    Rank* const out_tiebreakers,
-                    size_t* const out_num_tiebreakers) {
-    /* Validate input parameters */
-    if (cards == NULL || len != HAND_SIZE || out_tiebreakers == NULL || out_num_tiebreakers == NULL) {
-        return 0;
-    }
-
-    /* Use provided counts or compute locally */
-    int local_counts[RANK_ARRAY_SIZE];
-    const int* rank_count_array;
-
-    if (counts == NULL) {
-        rank_counts(cards, len, local_counts);
-        rank_count_array = local_counts;
-    } else {
-        rank_count_array = counts;
-    }
-
-    /* Check for trips or quads (count >= 3) - excludes full houses and quads */
-    for (int rank = RANK_TWO; rank <= RANK_ACE; rank++) {
-        if (rank_count_array[rank] >= 3) {
-            return 0;
-        }
-    }
-
-    /* Find exactly 2 ranks with count == 2 (pairs) */
-    Rank pairs[2];
-    int pair_count = 0;
-
-    for (int rank = RANK_TWO; rank <= RANK_ACE; rank++) {
-        if (rank_count_array[rank] == 2) {
-            if (pair_count < 2) {
-                pairs[pair_count++] = (Rank)rank;
-            }
-        }
-    }
-
-    /* Must have exactly 2 pairs */
-    if (pair_count != 2) {
-        return 0;
-    }
-
-    /* Sort pairs in descending order (high pair first) */
-    if (pairs[0] < pairs[1]) {
-        Rank temp = pairs[0];
-        pairs[0] = pairs[1];
-        pairs[1] = temp;
-    }
-
-    /* Find the kicker (count == 1) */
-    Rank kicker = 0;
-    for (int rank = RANK_TWO; rank <= RANK_ACE; rank++) {
-        if (rank_count_array[rank] == 1) {
-            kicker = (Rank)rank;
-            break;
-        }
-    }
-
-    /* Write tiebreakers: [high_pair, low_pair, kicker] */
-    out_tiebreakers[0] = pairs[0];  /* High pair */
-    out_tiebreakers[1] = pairs[1];  /* Low pair */
-    out_tiebreakers[2] = kicker;
-    *out_num_tiebreakers = 3;
-
-    return 1;
-}
-
-/**
- * @brief Detect one pair
- *
- * Detects if the hand contains exactly one pair (1 rank with 2 cards).
- * Returns tiebreakers in the format: [pair_rank, kicker1, kicker2, kicker3]
- * where kickers are sorted in descending order.
- *
- * The function validates all input parameters and handles the optional counts
- * parameter by computing counts internally if NULL is provided. It excludes
- * two pairs, trips, full houses, and quads.
- *
- * @param cards Array of exactly HAND_SIZE cards
- * @param len Must be HAND_SIZE
- * @param counts Optional pre-computed rank counts (can be NULL)
- * @param out_tiebreakers Output array for tiebreaker ranks
- * @param out_num_tiebreakers Pointer to receive count of tiebreakers
- * @return 1 if one pair, 0 otherwise
- */
-int detect_one_pair(const Card* const cards, const size_t len,
-                    const int* const counts,
-                    Rank* const out_tiebreakers,
-                    size_t* const out_num_tiebreakers) {
-    /* Validate input parameters */
-    if (cards == NULL || len != HAND_SIZE || out_tiebreakers == NULL || out_num_tiebreakers == NULL) {
-        return 0;
-    }
-
-    /* Use provided counts or compute locally */
-    int local_counts[RANK_ARRAY_SIZE];
-    const int* rank_count_array;
-
-    if (counts == NULL) {
-        rank_counts(cards, len, local_counts);
-        rank_count_array = local_counts;
-    } else {
-        rank_count_array = counts;
-    }
-
-    /* Check for trips or quads (count >= 3) - excludes full houses, trips, and quads */
-    for (int rank = RANK_TWO; rank <= RANK_ACE; rank++) {
-        if (rank_count_array[rank] >= 3) {
-            return 0;
-        }
-    }
-
-    /* Find exactly 1 rank with count == 2 (pair) */
-    Rank pair_rank = 0;
-    int pair_count = 0;
-
-    for (int rank = RANK_TWO; rank <= RANK_ACE; rank++) {
-        if (rank_count_array[rank] == 2) {
-            pair_rank = (Rank)rank;
-            pair_count++;
-        }
-    }
-
-    /* Must have exactly 1 pair */
-    if (pair_count != 1) {
-        return 0;
-    }
-
-    /* Find exactly 3 kickers (count == 1) */
-    Rank kickers[3];
-    int kicker_count = 0;
-
-    for (int rank = RANK_TWO; rank <= RANK_ACE; rank++) {
-        if (rank_count_array[rank] == 1) {
-            if (kicker_count < 3) {
-                kickers[kicker_count++] = (Rank)rank;
-            }
-        }
-    }
-
-    /* Should have exactly 3 kickers */
-    if (kicker_count != 3) {
-        return 0;
-    }
-
-    /* Sort kickers in descending order */
-    qsort(kickers, 3, sizeof(Rank), rank_compare_desc);
-
-    /* Write tiebreakers: [pair_rank, kicker1, kicker2, kicker3] */
-    out_tiebreakers[0] = pair_rank;
-    out_tiebreakers[1] = kickers[0];  /* Highest kicker */
-    out_tiebreakers[2] = kickers[1];  /* Middle kicker */
-    out_tiebreakers[3] = kickers[2];  /* Lowest kicker */
-    *out_num_tiebreakers = 4;
-
-    return 1;
-}
-
 int evaluate_hand(void) {
     /* Placeholder for hand evaluation */
     return 0;
-}
-
-/**
- * @brief Detect flush (non-straight)
- *
- * Detects if the hand is a flush (HAND_SIZE suited cards, non-sequential).
- * Uses is_flush() to verify all cards have the same suit, then uses
- * is_straight() to exclude straight flushes. Returns all HAND_SIZE ranks in
- * descending order as tiebreakers.
- *
- * @param cards Array of exactly HAND_SIZE cards
- * @param len Must be HAND_SIZE
- * @param out_tiebreakers Output array for tiebreaker ranks
- * @param out_num_tiebreakers Pointer to receive count of tiebreakers
- * @return 1 if flush, 0 otherwise
- */
-int detect_flush(const Card* const cards, const size_t len,
-                 Rank* const out_tiebreakers,
-                 size_t* const out_num_tiebreakers) {
-    /* Validate input parameters */
-    if (cards == NULL || len != HAND_SIZE || out_tiebreakers == NULL || out_num_tiebreakers == NULL) {
-        return 0;
-    }
-
-    /* Check if all cards are the same suit (flush) */
-    if (!is_flush(cards, len)) {
-        return 0;
-    }
-
-    /* Exclude straight flushes (return 0 if straight) */
-    if (is_straight(cards, len, NULL)) {
-        return 0;
-    }
-
-    /* Extract ranks into array */
-    Rank ranks[HAND_SIZE];
-    for (size_t i = 0; i < HAND_SIZE; i++) {
-        ranks[i] = (Rank)cards[i].rank;
-    }
-
-    /* Sort ranks in descending order */
-    qsort(ranks, HAND_SIZE, sizeof(Rank), rank_compare_desc);
-
-    /* Write all HAND_SIZE ranks to tiebreakers */
-    for (size_t i = 0; i < HAND_SIZE; i++) {
-        out_tiebreakers[i] = ranks[i];
-    }
-
-    /* Set number of tiebreakers to 5 */
-    *out_num_tiebreakers = HAND_SIZE;
-
-    return 1;
-}
-
-/**
- * @brief Detect straight (non-flush)
- *
- * Detects if the hand is a straight but NOT a flush (excludes straight flushes).
- * Uses is_flush() to reject flushes, then is_straight() to detect sequential ranks.
- * Returns the high card as a tiebreaker (RANK_FIVE for wheel straight).
- *
- * @param cards Array of exactly HAND_SIZE cards
- * @param len Must be HAND_SIZE
- * @param out_tiebreakers Output array for tiebreaker ranks
- * @param out_num_tiebreakers Pointer to receive count of tiebreakers
- * @return 1 if straight (non-flush), 0 otherwise
- */
-int detect_straight(const Card* const cards, const size_t len,
-                    Rank* const out_tiebreakers,
-                    size_t* const out_num_tiebreakers) {
-    /* Validate input parameters */
-    if (cards == NULL || len != HAND_SIZE || out_tiebreakers == NULL || out_num_tiebreakers == NULL) {
-        return 0;
-    }
-
-    /* Check if it's a flush - if so, return 0 (exclude straight flushes) */
-    if (is_flush(cards, len)) {
-        return 0;
-    }
-
-    /* Check if cards form a straight */
-    Rank high_card;
-    if (!is_straight(cards, len, &high_card)) {
-        return 0;
-    }
-
-    /* Write tiebreaker: [high_card] */
-    out_tiebreakers[0] = high_card;
-    *out_num_tiebreakers = 1;
-
-    return 1;
-}
-
-/**
- * @brief Detect high card (always succeeds for valid input)
- *
- * Detects high card, which is the fallback when no other hand category matches.
- * Returns all HAND_SIZE card ranks in descending order as tiebreakers. This function
- * always succeeds for valid HAND_SIZE-card input, making it the default/weakest hand.
- *
- * @param cards Array of exactly HAND_SIZE cards
- * @param len Must be HAND_SIZE
- * @param out_tiebreakers Output array for tiebreaker ranks
- * @param out_num_tiebreakers Pointer to receive count of tiebreakers
- * @return 1 if valid HAND_SIZE-card input, 0 otherwise
- */
-int detect_high_card(const Card* const cards, const size_t len,
-                     Rank* const out_tiebreakers,
-                     size_t* const out_num_tiebreakers) {
-    /* Validate input parameters */
-    if (cards == NULL || len != HAND_SIZE || out_tiebreakers == NULL || out_num_tiebreakers == NULL) {
-        return 0;
-    }
-
-    /* Extract ranks into array */
-    Rank ranks[HAND_SIZE];
-    for (size_t i = 0; i < HAND_SIZE; i++) {
-        ranks[i] = (Rank)cards[i].rank;
-    }
-
-    /* Sort ranks in descending order */
-    qsort(ranks, HAND_SIZE, sizeof(Rank), rank_compare_desc);
-
-    /* Write all HAND_SIZE ranks to tiebreakers */
-    for (size_t i = 0; i < HAND_SIZE; i++) {
-        out_tiebreakers[i] = ranks[i];
-    }
-
-    /* Set number of tiebreakers to 5 */
-    *out_num_tiebreakers = HAND_SIZE;
-
-    return 1;
 }
